feat: Add initial LaunchScreen.storyboard for iOS
This commit introduces the `LaunchScreen.storyboard` file to
the iOS project. This storyboard defines the initial launch screen
displayed when the app starts.

The launch screen includes a `ViewController` with a `imageView`
that displays the `LaunchImage`. The `imageView` is centered
within the main view. The background color is set to white.

This provides a standard launch screen experience for iOS users.