feat: Add initial Main.storyboard for Flutter integration
This commit introduces the `Main.storyboard` file, which is essential
for integrating the Flutter application into the native iOS environment.
It sets up a basic `FlutterViewController` as the initial view controller
within the storyboard. This allows the Flutter view to be displayed
correctly within the iOS application's user interface.  The background
color is set to white.