feat: Create Runner-Bridging-Header.h for Swift/ObjC interop
This commit introduces the `Runner-Bridging-Header.h` file, which
is essential for enabling communication between Swift and Objective-C
code within the Flutter iOS project.  Specifically, it imports
`GeneratedPluginRegistrant.h`. This allows Flutter plugins written
in Objective-C to be accessible from Swift code in the Runner app.
This header file is a standard requirement for mixed-language
Flutter iOS projects.