feat: Create initial generated plugin registrant for Linux
This commit adds the `generated_plugin_registrant.cc` file, which
provides a basic plugin registration mechanism for Flutter on Linux.

Currently, the `fl_register_plugins` function is empty. This serves
as a placeholder and will be populated with plugin registration calls
in subsequent commits as plugins are added to the project.

The file includes necessary headers and a comment indicating that it
should not be manually edited, as it is automatically generated.