feat: Add basic RunnerTests.swift file for macOS
This commit introduces the `RunnerTests.swift` file to the macOS
project. This file provides a basic XCTestCase class (`RunnerTests`)
with a placeholder `testExample` function. This serves as a starting
point for implementing unit and integration tests for the macOS
Flutter runner application. The initial test function includes a
comment directing developers to Apple's XCTest documentation.