feat: Add initial web/index.html for Flutter web app
This commit introduces the initial `index.html` file for the Flutter web application.
It includes basic HTML structure, meta tags for iOS and web app capabilities,
favicon and manifest links, and a placeholder for the base href that will be
replaced during the build process with the `--base-href` argument.  The
`flutter_bootstrap.js` script is also included for initializing the Flutter
web application.