feat: Add maskable 192x192 icon for progressive web app
This commit introduces a maskable 192x192 pixel icon
(Icon-maskable-192.png) to the web/icons directory. This
icon is designed for use as a maskable icon for Progressive
Web Apps (PWAs), allowing the operating system to apply a
mask to the icon to better integrate it with the device's UI.
The addition of this maskable icon improves the visual
consistency of the PWA on various platforms and devices.