refactor: Simplify UssdResponse.fromText default case
This commit refactors the `UssdResponse.fromText` method to simplify the default case where the input text does not contain a USSD command prefix.  Specifically, the creation of the `UssdResponse` object in the `else` block is made more concise by removing unnecessary intermediate variable assignments. This improves readability without changing the functionality. The `UssdResponse` is still created with the `textResponse` and `continueSession` set to `true`.