refactor: Improve HomeScreen BottomNavigationBar item declarations
This commit refactors the `HomeScreen` widget to improve the conciseness of the `BottomNavigationBarItem` declarations.

The change simplifies the item creation by directly assigning the `icon` and `label` properties within the `items` list, removing unnecessary line breaks and improving readability. This does not change the functionality of the BottomNavigationBar, only the way the items are defined.