refactor: Simplify conditional checks in build-release.yml
Simplifies the conditional checks for keystore and store password secrets in the build-release.yml workflow.

The conditions were updated from `secrets.KEYSTORE_BASE64 != ''` to `secrets.KEYSTORE_BASE64` and from `secrets.KEYSTORE_BASE64 == ''` to `!secrets.KEYSTORE_BASE64` for improved readability and conciseness. The conditions now evaluate to true if the secret exists and is not empty, or if it doesn't exists/is empty, respectively.  This affects keystore creation, key properties creation, APK building, and asset uploading jobs. The change aims to make the logic easier to understand and maintain.