feat: Add flutter_tts and speech_to_text dependencies
This commit adds the `flutter_tts` (v3.8.5) and `speech_to_text` (v6.6.0)
dependencies to the project. It also includes updates to transitive
dependencies such as `leak_tracker`, `leak_tracker_flutter_testing`,
`leak_tracker_testing`, `test_api`, and `vector_math` due to the
addition of the new packages and other dependency updates such as adding
`mockito` and `pedantic`. These changes are necessary to implement text-to-speech
and speech-to-text functionalities within the application.